@@ -27,119 +37,124 @@
   },
   "devDependencies": {
     "babel-eslint": "^7.1.1",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-transform-react-constant-elements": "^6.8.0",
-    "babel-plugin-transform-react-inline-elements": "^6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.2.6",
+    "babel-loader": "^7.1.5",
+    "babel-plugin-transform-react-constant-elements": "^6.23.0",
+    "babel-plugin-transform-react-inline-elements": "^6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.2.12",
     "babel-plugin-webpack-loaders": "^0.9.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.5.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-react-optimize": "^1.0.1",
-    "babel-preset-stage-0": "^6.5.0",
-    "chunk-manifest-webpack-plugin": "^1.1.2",
+    "babel-preset-stage-0": "^6.24.1",
+    "chunk-manifest-webpack-plugin": "github:catarak/chunk-manifest-webpack-plugin",
     "css-loader": "^0.23.1",
-    "cssnano": "^3.7.1",
-    "eslint": "^4.9.0",
+    "cssnano": "^3.10.0",
+    "eslint": "^4.19.1",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.12.3",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^2.0.0",
-    "node-sass": "^4.9.0",
-    "nodemon": "^1.9.2",
-    "postcss-cssnext": "^2.7.0",
+    "node-sass": "^4.11.0",
+    "nodemon": "^1.18.9",
+    "postcss-cssnext": "^2.11.0",
+    "jest": "^23.6.0",
     "postcss-focus": "^1.0.0",
     "postcss-loader": "^0.9.1",
-    "postcss-reporter": "^1.3.3",
-    "rimraf": "^2.6.2",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.13.1",
-    "webpack-manifest-plugin": "^2.0.0",
+    "postcss-reporter": "^1.4.1",
+    "rimraf": "^2.6.3",
+    "sass-loader": "^6.0.7",
+    "style-loader": "^0.13.2",
+    "webpack-manifest-plugin": "^2.0.4",
+    "react-test-renderer": "^16.6.0",
     "webpack-node-externals": "^1.7.2"
   },
   "engines": {
-    "node": ">=8.9.0"
+    "node": "10.15.0",
+    "npm": "6.4.1"
   },
   "dependencies": {
     "archiver": "^1.1.0",
-    "async": "^2.0.0",
+    "async": "^2.6.1",
     "axios": "^0.12.0",
-    "babel-core": "^6.26.0",
-    "babel-polyfill": "^6.8.0",
-    "babel-register": "^6.8.0",
+    "babel-core": "^6.26.3",
+    "babel-polyfill": "^6.26.0",
+    "babel-register": "^6.26.0",
     "bcrypt-nodejs": "0.0.3",
     "blob-util": "^1.2.1",
-    "body-parser": "^1.15.1",
-    "bson-objectid": "^1.1.4",
-    "classnames": "^2.2.5",
+    "body-parser": "^1.18.3",
+    "bson-objectid": "^1.2.4",
+    "classnames": "^2.2.6",
     "clipboard": "^1.7.1",
-    "codemirror": "^5.38.0",
-    "connect-mongo": "^1.2.0",
-    "console-feed": "^2.8.1",
-    "cookie-parser": "^1.4.1",
-    "cors": "^2.8.1",
-    "cross-env": "^5.1.3",
+    "codemirror": "^5.42.2",
+    "connect-mongo": "^1.3.2",
+    "console-feed": "^2.8.5",
+    "cookie-parser": "^1.4.3",
+    "cors": "^2.8.5",
+    "cross-env": "^5.2.0",
     "csslint": "^0.10.0",
+    "date-fns": "^1.30.1",
     "decomment": "^0.8.7",
     "dotenv": "^2.0.0",
     "dropzone": "^4.3.0",
     "escape-string-regexp": "^1.0.5",
     "eslint-loader": "^1.3.0",
-    "express": "^4.13.4",
-    "express-basic-auth": "^1.1.5",
-    "express-session": "^1.13.0",
-    "htmlhint": "^0.9.13",
-    "is-url": "^1.2.2",
-    "js-beautify": "^1.6.4",
+    "express": "^4.16.4",
+    "express-basic-auth": "^1.1.6",
+    "express-session": "^1.15.6",
+    "htmlhint": "^0.10.1",
+    "is-url": "^1.2.4",
+    "js-beautify": "^1.8.9",
     "jsdom": "^9.8.3",
-    "jshint": "^2.9.4",
-    "lodash": "^4.16.4",
+    "jshint": "^2.10.1",
+    "lodash": "^4.17.11",
     "loop-protect": "github:catarak/loop-protect",
     "mjml": "^3.3.2",
-    "moment": "^2.14.1",
     "mongoose": "^4.6.8",
     "node-uuid": "^1.4.7",
     "nodemailer": "^2.6.4",
-    "nodemailer-mailgun-transport": "^1.2.2",
+    "nodemailer-mailgun-transport": "^1.4.0",
     "passport": "^0.3.2",
     "passport-github": "^1.1.0",
     "passport-google-oauth20": "^1.0.0",
     "passport-local": "^1.0.0",
     "pretty-bytes": "^3.0.1",
-    "primer-tooltips": "^1.4.1",
-    "project-name-generator": "^2.1.3",
-    "prop-types": "^15.6.0",
+    "primer-tooltips": "^1.5.11",
+    "project-name-generator": "^2.1.5",
+    "prop-types": "^15.6.2",
     "q": "^1.4.1",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-helmet": "^5.1.3",
-    "react-hot-loader": "^4.1.2",
+    "react-hot-loader": "^4.6.3",
     "react-inlinesvg": "^0.7.5",
-    "react-redux": "^5.0.6",
-    "react-router": "^3.2.0",
-    "react-split-pane": "^0.1.44",
-    "react-tabs": "^2.2.1",
-    "redux": "^3.5.2",
-    "redux-devtools": "^3.4.1",
+    "react-redux": "^5.1.1",
+    "react-router": "^3.2.1",
+    "react-split-pane": "^0.1.85",
+    "react-tabs": "^2.3.0",
+    "redux": "^3.7.2",
+    "redux-devtools": "^3.4.2",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-log-monitor": "^1.4.0",
     "redux-form": "^5.3.3",
-    "redux-thunk": "^2.1.0",
-    "request": "^2.76.0",
+    "redux-thunk": "^2.3.0",
+    "request": "^2.88.0",
     "request-promise": "^4.1.1",
     "s3": "^4.4.0",
     "s3-policy": "^0.2.0",
     "sass-extract": "^2.1.0",
     "sass-extract-js": "^0.4.0",
     "sass-extract-loader": "^1.1.0",
-    "shortid": "^2.2.6",
-    "slugify": "^1.2.9",
+    "shortid": "^2.2.14",
+    "slugify": "^1.3.4",
     "srcdoc-polyfill": "^0.2.0",
     "url": "^0.11.0",
-    "webpack": "^3.1.0",
+    "webpack": "^3.12.0",
     "webpack-dev-middleware": "^2.0.6",
-    "webpack-hot-middleware": "^2.10.0",
-    "xhr": "^2.2.1"
+    "webpack-hot-middleware": "^2.24.3",
+    "xhr": "^2.5.0"
   }
 }
