@@ -2,13 +2,15 @@ import { browserHistory } from 'react-router';
 import axios from 'axios';
 import objectID from 'bson-objectid';
 import each from 'async/each';
-import { isEqual } from 'lodash';
+import { isEqual, pick } from 'lodash';
 import * as ActionTypes from '../../../constants';
 import { showToast, setToastText } from './toast';
-import { setUnsavedChanges,
+import {
+  setUnsavedChanges,
   justOpenedProject,
   resetJustOpenedProject,
-  showErrorModal } from './ide';
+  showErrorModal
+} from './ide';
 import { clearState, saveState } from '../../../persistState';
 
 const __process = (typeof global !== 'undefined' ? global : window).process;
@@ -64,7 +66,7 @@ export function clearPersistedState() {
   };
 }
 
-export function saveProject(autosave = false) {
+export function saveProject(selectedFile = null, autosave = false) {
   return (dispatch, getState) => {
     const state = getState();
     if (state.user.id && state.project.owner && state.project.owner.id !== state.user.id) {
@@ -72,12 +74,20 @@ export function saveProject(autosave = false) {
     }
     const formParams = Object.assign({}, state.project);
     formParams.files = [...state.files];
+    if (selectedFile) {
+      console.log('selected file being updated');
+      const fileToUpdate = formParams.files.find(file => file.id === selectedFile.id);
+      fileToUpdate.content = selectedFile.content;
+    }
     if (state.project.id) {
       return axios.put(`${ROOT_URL}/projects/${state.project.id}`, formParams, { withCredentials: true })
         .then((response) => {
           const currentState = getState();
           const savedProject = Object.assign({}, response.data);
-          if (!isEqual(currentState.files, response.data.files)) {
+          if (!isEqual(
+            pick(currentState.files, ['name', 'children', 'content']),
+            pick(response.data.files, ['name', 'children', 'content'])
+          )) {
             savedProject.files = currentState.files;
             dispatch(setUnsavedChanges(true));
           } else {
@@ -151,7 +161,7 @@ export function saveProject(autosave = false) {
 
 export function autosaveProject() {
   return (dispatch, getState) => {
-    saveProject(true)(dispatch, getState);
+    saveProject(null, true)(dispatch, getState);
   };
 }
 
