@@ -1,6 +1,10 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import Clipboard from 'clipboard';
+import InlineSVG from 'react-inlinesvg';
+import classNames from 'classnames';
+
+import shareUrl from '../../../images/share.svg';
 
 class CopyableInput extends React.Component {
   constructor(props) {
@@ -34,8 +38,12 @@ class CopyableInput extends React.Component {
       value,
       hasPreviewLink
     } = this.props;
+    const copyableInputClass = classNames({
+      'copyable-input': true,
+      'copyable-input--with-preview': hasPreviewLink
+    });
     return (
-      <div className="copyable-input">
+      <div className={copyableInputClass}>
         <div
           className="copyable-input__value-container tooltipped-no-delay"
           aria-label="Copied to Clipboard!"
@@ -44,10 +52,7 @@ class CopyableInput extends React.Component {
         >
           <label className="copyable-input__label" htmlFor={`copyable-input__value-${label}`}>
             <div className="copyable-input__label-container">
-              {label} {hasPreviewLink &&
-                <a target="_blank" href={value}>
-                  Open
-                </a>}
+              {label}
             </div>
             <input
               type="text"
@@ -59,6 +64,17 @@ class CopyableInput extends React.Component {
             />
           </label>
         </div>
+        {hasPreviewLink &&
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href={value}
+            className="copyable-input__preview"
+            title={`open ${label.toLowerCase()} view in new tab`}
+          >
+            <InlineSVG src={shareUrl} alt={`open ${label} view in new tab`} />
+          </a>
+        }
       </div>
     );
   }
