@@ -15,8 +15,8 @@ class Preferences extends React.Component {
   constructor(props) {
     super(props);
     this.handleUpdateAutosave = this.handleUpdateAutosave.bind(this);
+    this.handleUpdateLinewrap = this.handleUpdateLinewrap.bind(this);
     this.handleUpdateFont = this.handleUpdateFont.bind(this);
-    this.handleUpdateIndentation = this.handleUpdateIndentation.bind(this);
     this.handleLintWarning = this.handleLintWarning.bind(this);
   }
 
@@ -34,23 +34,14 @@ class Preferences extends React.Component {
     this.props.setFontSize(value);
   }
 
-  handleUpdateIndentation(event) {
-    let value = parseInt(event.target.value, 10);
-    if (Number.isNaN(value)) {
-      value = 2;
-    }
-    if (value > 6) {
-      value = 6;
-    }
-    if (value < 0) {
-      value = 0;
-    }
-    this.props.setIndentation(value);
-  }
-
   handleUpdateAutosave(event) {
     const value = event.target.value === 'true';
     this.props.setAutosave(value);
+  }
+
+  handleUpdateLinewrap(event) {
+    const value = event.target.value === 'true';
+    this.props.setLinewrap(value);
   }
 
   handleLintWarning(event) {
@@ -145,65 +136,6 @@ class Preferences extends React.Component {
               </button>
             </div>
             <div className="preference">
-              <h4 className="preference__title">Indentation amount</h4>
-              <button
-                className="preference__minus-button"
-                onClick={() => this.props.setIndentation(this.props.indentationAmount - 2)}
-                aria-label="decrease indentation amount"
-                disabled={this.props.indentationAmount <= 0}
-              >
-                <InlineSVG src={minusUrl} alt="DecreaseIndentation Amount" />
-                <h6 className="preference__label">Decrease</h6>
-              </button>
-              <input
-                className="preference__value"
-                aria-live="polite"
-                aria-atomic="true"
-                value={this.props.indentationAmount}
-                onChange={this.handleUpdateIndentation}
-                ref={(element) => { this.indentationInput = element; }}
-                onClick={() => {
-                  this.indentationInput.select();
-                }}
-              />
-              <button
-                className="preference__plus-button"
-                onClick={() => this.props.setIndentation(this.props.indentationAmount + 2)}
-                aria-label="increase indentation amount"
-                disabled={this.props.indentationAmount >= 6}
-              >
-                <InlineSVG src={plusUrl} alt="IncreaseIndentation Amount" />
-                <h6 className="preference__label">Increase</h6>
-              </button>
-              <input
-                type="radio"
-                onChange={this.props.indentWithSpace}
-                aria-label="indentation with space"
-                name="indentation"
-                id="indentation-space"
-                className="preference__radio-button"
-                value="Spaces"
-                checked={!this.props.isTabIndent}
-              />
-              <label
-                htmlFor="indentation-space"
-                className="preference__option preference__whitespace-button"
-              >
-                Spaces
-              </label>
-              <input
-                type="radio"
-                onChange={this.props.indentWithTab}
-                aria-label="indentation with tab"
-                name="indentation"
-                id="indentation-tab"
-                className="preference__radio-button"
-                value="Tabs"
-                checked={this.props.isTabIndent}
-              />
-              <label htmlFor="indentation-tab" className="preference__option preference__whitespace-button">Tabs</label>
-            </div>
-            <div className="preference">
               <h4 className="preference__title">Autosave</h4>
               <div className="preference__options">
                 <input
@@ -228,6 +160,33 @@ class Preferences extends React.Component {
                   checked={!this.props.autosave}
                 />
                 <label htmlFor="autosave-off" className="preference__option">Off</label>
+              </div>
+            </div>
+            <div className="preference">
+              <h4 className="preference__title">Word Wrap</h4>
+              <div className="preference__options">
+                <input
+                  type="radio"
+                  onChange={() => this.props.setLinewrap(true)}
+                  aria-label="linewrap on"
+                  name="linewrap"
+                  id="linewrap-on"
+                  className="preference__radio-button"
+                  value="On"
+                  checked={this.props.linewrap}
+                />
+                <label htmlFor="linewrap-on" className="preference__option">On</label>
+                <input
+                  type="radio"
+                  onChange={() => this.props.setLinewrap(false)}
+                  aria-label="linewrap off"
+                  name="linewrap"
+                  id="linewrap-off"
+                  className="preference__radio-button"
+                  value="Off"
+                  checked={!this.props.linewrap}
+                />
+                <label htmlFor="linewrap-off" className="preference__option">Off</label>
               </div>
             </div>
           </TabPanel>
@@ -318,14 +277,11 @@ class Preferences extends React.Component {
 
 Preferences.propTypes = {
   fontSize: PropTypes.number.isRequired,
-  indentationAmount: PropTypes.number.isRequired,
-  setIndentation: PropTypes.func.isRequired,
-  indentWithSpace: PropTypes.func.isRequired,
-  indentWithTab: PropTypes.func.isRequired,
-  isTabIndent: PropTypes.bool.isRequired,
   setFontSize: PropTypes.func.isRequired,
   autosave: PropTypes.bool.isRequired,
+  linewrap: PropTypes.bool.isRequired,
   setAutosave: PropTypes.func.isRequired,
+  setLinewrap: PropTypes.func.isRequired,
   textOutput: PropTypes.bool.isRequired,
   gridOutput: PropTypes.bool.isRequired,
   soundOutput: PropTypes.bool.isRequired,
