@@ -12,6 +12,25 @@ class Sidebar extends React.Component {
     super(props);
     this.resetSelectedFile = this.resetSelectedFile.bind(this);
     this.toggleProjectOptions = this.toggleProjectOptions.bind(this);
+    this.onBlurComponent = this.onBlurComponent.bind(this);
+    this.onFocusComponent = this.onFocusComponent.bind(this);
+
+    this.state = {
+      isFocused: false,
+    };
+  }
+
+  onBlurComponent() {
+    this.setState({ isFocused: false });
+    setTimeout(() => {
+      if (!this.state.isFocused) {
+        this.props.closeProjectOptions();
+      }
+    }, 200);
+  }
+
+  onFocusComponent() {
+    this.setState({ isFocused: true });
   }
 
   resetSelectedFile() {
@@ -41,11 +60,12 @@ class Sidebar extends React.Component {
   }
 
   render() {
+    const canEditProject = this.userCanEditProject();
     const sidebarClass = classNames({
       'sidebar': true,
       'sidebar--contracted': !this.props.isExpanded,
       'sidebar--project-options': this.props.projectOptionsVisible,
-      'sidebar--cant-edit': !this.userCanEditProject()
+      'sidebar--cant-edit': !canEditProject
     });
 
     return (
@@ -64,25 +84,45 @@ class Sidebar extends React.Component {
               tabIndex="0"
               ref={(element) => { this.sidebarOptions = element; }}
               onClick={this.toggleProjectOptions}
-              onBlur={() => setTimeout(this.props.closeProjectOptions, 200)}
+              onBlur={this.onBlurComponent}
+              onFocus={this.onFocusComponent}
             >
               <InlineSVG src={downArrowUrl} />
             </button>
             <ul className="sidebar__project-options">
               <li>
-                <button aria-label="add folder" onClick={this.props.newFolder} >
+                <button
+                  aria-label="add folder"
+                  onClick={() => {
+                    this.props.newFolder();
+                    setTimeout(this.props.closeProjectOptions, 0);
+                  }}
+                  onBlur={this.onBlurComponent}
+                  onFocus={this.onFocusComponent}
+                >
                   Add folder
                 </button>
               </li>
               <li>
-                <button aria-label="add file" onClick={this.props.newFile} >
+                <button
+                  aria-label="add file"
+                  onClick={() => {
+                    this.props.newFile();
+                    setTimeout(this.props.closeProjectOptions, 0);
+                  }}
+                  onBlur={this.onBlurComponent}
+                  onFocus={this.onFocusComponent}
+                >
                   Add file
                 </button>
               </li>
             </ul>
           </div>
         </div>
-        <ConnectedFileNode id={this.props.files.filter(file => file.name === 'root')[0].id} />
+        <ConnectedFileNode
+          id={this.props.files.filter(file => file.name === 'root')[0].id}
+          canEdit={canEditProject}
+        />
       </nav>
     );
   }
