@@ -1,7 +1,8 @@
+import classNames from 'classnames';
+import differenceInMilliseconds from 'date-fns/difference_in_milliseconds';
+import distanceInWordsToNow from 'date-fns/distance_in_words_to_now';
 import PropTypes from 'prop-types';
 import React from 'react';
-import moment from 'moment';
-import classNames from 'classnames';
 
 class Timer extends React.Component {
   constructor(props) {
@@ -20,16 +21,19 @@ class Timer extends React.Component {
   }
 
   showSavedTime() {
-    if (Math.abs(moment().diff(this.props.projectSavedTime)) < 10000) {
+    const now = new Date();
+    if (Math.abs(differenceInMilliseconds(now, this.props.projectSavedTime) < 10000)) {
       return 'Saved: just now';
-    } else if (Math.abs(moment().diff(this.props.projectSavedTime)) < 20000) {
+    } else if (differenceInMilliseconds(now, this.props.projectSavedTime) < 20000) {
       return 'Saved: 15 seconds ago';
-    } else if (Math.abs(moment().diff(this.props.projectSavedTime)) < 30000) {
+    } else if (differenceInMilliseconds(now, this.props.projectSavedTime) < 30000) {
       return 'Saved: 25 seconds ago';
-    } else if (Math.abs(moment().diff(this.props.projectSavedTime)) < 46000) {
+    } else if (differenceInMilliseconds(now, this.props.projectSavedTime) < 46000) {
       return 'Saved: 35 seconds ago';
     }
-    return `Saved: ${moment(this.props.projectSavedTime).fromNow()}`;
+    return `Saved: ${distanceInWordsToNow(this.props.projectSavedTime, {
+      includeSeconds: true
+    })} ago`;
   }
 
   render() {
