@@ -4,7 +4,8 @@ import {
   injectMediaUrls,
   resolvePathsForElementsWithAttribute,
   resolveScripts,
-  resolveStyles } from '../utils/previewGeneration';
+  resolveStyles
+} from '../utils/previewGeneration';
 import { get404Sketch } from '../views/404Page';
 
 export function serveProject(req, res) {
