@@ -4,14 +4,17 @@ import slugify from 'slugify';
 
 const { Schema } = mongoose;
 
-const fileSchema = new Schema({
-  name: { type: String, default: 'sketch.js' },
-  content: { type: String, default: '' },
-  url: { type: String },
-  children: { type: [String], default: [] },
-  fileType: { type: String, default: 'file' },
-  isSelectedFile: { type: Boolean }
-}, { timestamps: true, _id: true, usePushEach: true });
+const fileSchema = new Schema(
+  {
+    name: { type: String, default: 'sketch.js' },
+    content: { type: String, default: '' },
+    url: { type: String },
+    children: { type: [String], default: [] },
+    fileType: { type: String, default: 'file' },
+    isSelectedFile: { type: Boolean }
+  },
+  { timestamps: true, _id: true, usePushEach: true }
+);
 
 fileSchema.virtual('id').get(function getFileId() {
   return this._id.toHexString();
@@ -21,14 +24,17 @@ fileSchema.set('toJSON', {
   virtuals: true
 });
 
-const projectSchema = new Schema({
-  name: { type: String, default: "Hello p5.js, it's the server" },
-  user: { type: Schema.Types.ObjectId, ref: 'User' },
-  serveSecure: { type: Boolean, default: false },
-  files: { type: [fileSchema] },
-  _id: { type: String, default: shortid.generate },
-  slug: { type: String }
-}, { timestamps: true, usePushEach: true });
+const projectSchema = new Schema(
+  {
+    name: { type: String, default: "Hello p5.js, it's the server" },
+    user: { type: Schema.Types.ObjectId, ref: 'User' },
+    serveSecure: { type: Boolean, default: false },
+    files: { type: [fileSchema] },
+    _id: { type: String, default: shortid.generate },
+    slug: { type: String }
+  },
+  { timestamps: true, usePushEach: true }
+);
 
 projectSchema.virtual('id').get(function getProjectId() {
   return this._id;
